backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
transformers>=4.35.0
torch>=2.2.0
numpy>=1.24.3
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.5.0
wikipedia>=1.4.0

